app/backend/requirements.txt
# Leonida Backend — Railway-ready dependencies
# Core
fastapi==0.110.1
uvicorn==0.25.0
starlette==0.37.2
pydantic==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.24

# Database
motor==3.3.1
pymongo==4.5.0
dnspython==2.8.0

# Auth
bcrypt==4.1.3
passlib==1.7.4
python-jose==3.5.0
PyJWT==2.12.1
cryptography==46.0.7

# News scraper
feedparser==6.0.12
httpx==0.28.1
APScheduler==3.11.2

# Email notifications
requests==2.33.1
resend

# Utilities
certifi==2026.2.25
